refactor: add layers for layered architecture and move files into their corresponding layer
