+++ src/domain/store-errors.ts
