+++ src/domain/user-store-errors.ts
