+++ src/utils/utils.ts
@@ -1,7 +1,7 @@
 // validate username doesn't exist already
 import { PrismaClient } from '@prisma/client'
 import * as jwt from 'jsonwebtoken'
-import {userModel} from '../domain/stores/user-store'
+import {userModel} from '../infra/stores/user-store'
 
 const prisma = new PrismaClient()
